+++ Samples~/Examples/Scripts/InheritedExample.cs
@@ -1,0 +1,9 @@
+﻿using UnityEngine;
+
+namespace YellowSquad.AssetPath.Samples.Examples
+{
+    public class InheritedExample : Example
+    {
+        [SerializeField] private ResourcesReference<Rigidbody> _rigidbodyPath;
+    }
+}